#2696 - Document-level recommendations
- Continue moving extraction logic into the suggestion supports
