+++ inception/inception-recommendation/src/main/java/de/tudarmstadt/ukp/inception/recommendation/relation/RelationSuggestionSupport.java
@@ -62,8 +62,7 @@ import de.tudarmstadt.ukp.inception.recommendation.api.model.Position;
 import de.tudarmstadt.ukp.inception.recommendation.api.model.RelationPosition;
 import de.tudarmstadt.ukp.inception.recommendation.api.model.RelationSuggestion;
 import de.tudarmstadt.ukp.inception.recommendation.api.model.SuggestionGroup;
-import de.tudarmstadt.ukp.inception.recommendation.service.ExtractionContext;
-import de.tudarmstadt.ukp.inception.recommendation.service.SuggestionExtraction;
+import de.tudarmstadt.ukp.inception.recommendation.api.recommender.ExtractionContext;
 import de.tudarmstadt.ukp.inception.schema.api.AnnotationSchemaService;
 import de.tudarmstadt.ukp.inception.schema.api.adapter.AnnotationComparisonUtils;
 import de.tudarmstadt.ukp.inception.schema.api.adapter.AnnotationException;
@@ -380,10 +379,8 @@ public class RelationSuggestionSupport
 
     public static void extractSuggestion(ExtractionContext ctx, TOP predictedFS)
     {
-        var autoAcceptMode = SuggestionExtraction.getAutoAcceptMode(predictedFS,
-                ctx.getModeFeature());
-        var labels = SuggestionExtraction.getPredictedLabels(predictedFS, ctx.getLabelFeature(),
-                ctx.isMultiLabels());
+        var autoAcceptMode = getAutoAcceptMode(predictedFS, ctx.getModeFeature());
+        var labels = getPredictedLabels(predictedFS, ctx.getLabelFeature(), ctx.isMultiLabels());
         var score = predictedFS.getDoubleValue(ctx.getScoreFeature());
         var scoreExplanation = predictedFS.getStringValue(ctx.getScoreExplanationFeature());
